requirements.txt
python-telegram-bot==20.3
mysql-connector-python==8.0.32
python-dotenv==1.0.0
openpyxl==3.1.2
pytest==7.4.0
python-dateutil==2.8.2